+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Use ES2015 modules instead of UMD.
 - #1257 Prefer 'probably' over 'maybe' when evaluating audio play support.
 - #1261 File upload not working
+- #1252 Correctly reflect the state in bookmark icon title.
 
 ## 4.0.3 (2018-10-22)